requirements.txt
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.9.4
Faker==2.0.2
gunicorn==19.9.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
